Moved translation helpers to sub namespace
