+++ config/app.php
@@ -17,7 +17,7 @@ return [
         \Engelsystem\Database\DatabaseServiceProvider::class,
         \Engelsystem\Http\RequestServiceProvider::class,
         \Engelsystem\Http\SessionServiceProvider::class,
-        \Engelsystem\Helpers\TranslationServiceProvider::class,
+        \Engelsystem\Helpers\Translation\TranslationServiceProvider::class,
         \Engelsystem\Http\ResponseServiceProvider::class,
         \Engelsystem\Http\Psr7ServiceProvider::class,
         \Engelsystem\Helpers\AuthenticatorServiceProvider::class,
