+++ src/Middleware/LegacyMiddleware.php
@@ -3,7 +3,7 @@
 namespace Engelsystem\Middleware;
 
 use Engelsystem\Helpers\Authenticator;
-use Engelsystem\Helpers\Translator;
+use Engelsystem\Helpers\Translation\Translator;
 use Engelsystem\Http\Request;
 use Engelsystem\Http\Response;
 use Psr\Container\ContainerInterface;
