+++ src/Middleware/SetLocale.php
@@ -2,7 +2,7 @@
 
 namespace Engelsystem\Middleware;
 
-use Engelsystem\Helpers\Translator;
+use Engelsystem\Helpers\Translation\Translator;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
