+++ tests/Unit/Helpers/Translation/TranslationServiceProviderTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Engelsystem\Test\Unit\Helpers;
+namespace Engelsystem\Test\Unit\Helpers\Translation;
 
 use Engelsystem\Config\Config;
-use Engelsystem\Helpers\TranslationServiceProvider;
-use Engelsystem\Helpers\Translator;
+use Engelsystem\Helpers\Translation\TranslationServiceProvider;
+use Engelsystem\Helpers\Translation\Translator;
 use Engelsystem\Test\Unit\ServiceProviderTest;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Component\HttpFoundation\Session\Session;
@@ -12,7 +12,7 @@ use Symfony\Component\HttpFoundation\Session\Session;
 class TranslationServiceProviderTest extends ServiceProviderTest
 {
     /**
-     * @covers \Engelsystem\Helpers\TranslationServiceProvider::register()
+     * @covers \Engelsystem\Helpers\Translation\TranslationServiceProvider::register()
      */
     public function testRegister()
     {
