+++ tests/Unit/HelpersTest.php
@@ -6,7 +6,7 @@ use Engelsystem\Application;
 use Engelsystem\Config\Config;
 use Engelsystem\Container\Container;
 use Engelsystem\Helpers\Authenticator;
-use Engelsystem\Helpers\Translator;
+use Engelsystem\Helpers\Translation\Translator;
 use Engelsystem\Http\Request;
 use Engelsystem\Http\Response;
 use Engelsystem\Http\UrlGeneratorInterface;
